management-ui/backend/requirements.txt
docker>=7.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
